Add Prompt Component
This adds a Yes/No Prompt component. This shows your prompt plus a "Yes"
and "No" buttons. The user can select these buttons using the left and
right arrow keys. By default and for safety, "No" is selected.

This is meant to be added to an if/else statement as it returns a
boolean based on the option selected.