Refactor terminal size usage
This stops subtracting from the rows and columns that we get from the
terminal size package and just uses the raw values making widgets do
their own adjustments.

This also stops drawing a box around viewport.

This also moves Row and dimensions to the main file, and adds a eols()
function for printing multiple EOL chars.