Add StylePersist()
This adds a persistent version of Style() that prints the styles with no
reset instead of returning the text with the styles prepended.

This allows for Styles you would like to persist for an entire line for
example.

As such, this adds a Reset constant to append at the end yourself.