Improve Viewport
Call Bell function when we reach the end and the beginning of the
context we're scrolling.

I wanted to add page up and page down so I moved move up and move down
logic into functions, key presses for viewport will just give a smaller
or greater value of how much to move up or down.